Gulp task added to create a crx file for chrome extension.
