--- plugins/chrome/manifest.json
@@ -1,15 +1,0 @@
-{
-	"manifest_version":	2,
-	"name":	"KodeBeagle",
-	"description":	"KodeBeagle Chrome extension. An extension that makes searching selected piece of java code easily searcheable on KodeBeagle.
- Currently KodeBeagle Supports only JAVA programming language. It does some optimizations like selecting keywords from
- selected code automatically by filtering 'java.lang.*' and some keywords.",
-	"version":	"0.1",
-	"background":	{
-		"scripts":	["background.js"]
-	},
-	"permissions":	[
-		"contextMenus"
-	]
-
-}
