+++ plugins/chrome/src/java_lang_pkg.json
@@ -1,0 +1,115 @@
+{
+  "appendable": true,
+  "autocloseable": true,
+  "charsequence": true,
+  "cloneable": true,
+  "comparable": true,
+  "iterable": true,
+  "readable": true,
+  "runnable": true,
+  "thread.uncaughtexceptionhandler": true,
+
+  "boolean": true,
+  "byte": true,
+  "character": true,
+  "character.subset": true,
+  "character.unicodeblock": true,
+  "class": true,
+  "classloader": true,
+  "classvalue": true,
+  "compiler": true,
+  "double": true,
+  "enum": true,
+  "float": true,
+  "inheritablethreadlocal": true,
+  "integer": true,
+  "long": true,
+  "math": true,
+  "number": true,
+  "object": true,
+  "package": true,
+  "process": true,
+  "processbuilder": true,
+  "processbuilder.redirect": true,
+  "runtime": true,
+  "runtimepermission": true,
+  "securitymanager": true,
+  "short": true,
+  "stacktraceelement": true,
+  "strictmath": true,
+  "string": true,
+  "stringbuffer": true,
+  "stringbuilder": true,
+  "system": true,
+  "thread": true,
+  "threadgroup": true,
+  "threadlocal": true,
+  "throwable": true,
+  "void": true,
+
+  "character.unicodescript": true,
+  "processbuilder.redirect.type": true,
+  "thread.state": true,
+
+  "arithmeticexception": true,
+  "arrayindexoutofboundsexception": true,
+  "arraystoreexception": true,
+  "classcastexception": true,
+  "classnotfoundexception": true,
+  "clonenotsupportedexception": true,
+  "enumconstantnotpresentexception": true,
+  "exception": true,
+  "illegalaccessexception": true,
+  "illegalargumentexception": true,
+  "illegalmonitorstateexception": true,
+  "illegalstateexception": true,
+  "illegalthreadstateexception": true,
+  "indexoutofboundsexception": true,
+  "instantiationexception": true,
+  "interruptedexception": true,
+  "negativearraysizeexception": true,
+  "nosuchfieldexception": true,
+  "nosuchmethodexception": true,
+  "nullpointerexception": true,
+  "numberformatexception": true,
+  "reflectiveoperationexception": true,
+  "runtimeexception": true,
+  "securityexception": true,
+  "stringindexoutofboundsexception": true,
+  "typenotpresentexception": true,
+  "unsupportedoperationexception": true,
+
+  "abstractmethoderror": true,
+  "assertionerror": true,
+  "bootstrapmethoderror": true,
+  "classcircularityerror": true,
+  "classformaterror": true,
+  "error": true,
+  "exceptionininitializererror": true,
+  "illegalaccesserror": true,
+  "incompatibleclasschangeerror": true,
+  "instantiationerror": true,
+  "internalerror": true,
+  "linkageerror": true,
+  "noclassdeffounderror": true,
+  "nosuchfielderror": true,
+  "nosuchmethoderror": true,
+  "outofmemoryerror": true,
+  "stackoverflowerror": true,
+  "threaddeath": true,
+  "unknownerror": true,
+  "unsatisfiedlinkerror": true,
+  "unsupportedclassversioneerror": true,
+  "verifyerror": true,
+  "virtualmachineerror": true,
+
+  "deprecated": true,
+  "override": true,
+  "safevarargs": true,
+  "suppresswarnings": true,
+
+  "int": true,
+  "char": true,
+  "long": true
+
+}
